Added CORS error handling to custom-login sample (for login and logout) Moved CORS error modal to separate component
