+++ custom-login/src/Login.jsx
@@ -16,7 +16,7 @@ import '@okta/okta-signin-widget/dist/css/okta-sign-in.min.css';
 
 import config from './config';
 
-const Login = () => {
+const Login = ({ setCorsErrorModalOpen }) => {
   const { oktaAuth } = useOktaAuth();
   const widgetRef = useRef();
 
@@ -52,13 +52,21 @@ const Login = () => {
     widget.renderEl(
       { el: widgetRef.current },
       (res) => {
-        console.log(res);
         oktaAuth.handleLoginRedirect(res.tokens);
       },
       (err) => {
         throw err;
       },
     );
+
+    // Note: Can't distinguish CORS error from other network errors
+    const isCorsError = (err) => (err.name === 'AuthApiError' && !err.statusCode);
+
+    widget.on('afterError', (_context, error) => {
+      if (isCorsError(error)) {
+        setCorsErrorModalOpen(true);
+      }
+    });
 
     return () => widget.remove();
   }, [oktaAuth]);
