+++ okta-hosted-login/src/Navbar.jsx
@@ -13,26 +13,13 @@
 import { useOktaAuth } from '@okta/okta-react';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { Container, Icon, Image, Menu, Modal } from 'semantic-ui-react';
-import config from './config';
+import { Container, Icon, Image, Menu } from 'semantic-ui-react';
 
-const Navbar = () => {
-  const [errorModalOpen, setErrorModalOpen] = React.useState(false);
+const Navbar = ({ setCorsErrorModalOpen }) => {
   const { authState, oktaAuth } = useOktaAuth();
 
   // Note: Can't distinguish CORS error from other network errors
   const isCorsError = (err) => (err.name === 'AuthApiError' && !err.errorCode && err.xhr.message === 'Failed to fetch');
-
-  // Build URL to "Trusted Origins" page in Admin panel
-  const { issuer } = config.oidc;
-  const baseUrl = issuer.split('/oauth2')[0];
-  const hostParts = new URL(baseUrl).host.split('.');
-  hostParts[0] += '-admin';
-  const adminHost = hostParts.join('.');
-  const corsAdminUrl = `https://${adminHost}/admin/access/api/trusted_origins`;
-
-  // URL to guide for enabling CORS
-  const guideUrl = 'https://developer.okta.com/docs/guides/enable-cors/granting-cors/';
 
   const login = async () => oktaAuth.signInWithRedirect();
 
@@ -41,44 +28,12 @@ const Navbar = () => {
       await oktaAuth.signOut();
     } catch (err) {
       if (isCorsError(err)) {
-        setErrorModalOpen(true);
+        setCorsErrorModalOpen(true);
       } else {
         throw err;
       }
     }
   };
-
-  // CORS error modal
-  const errorModal = (
-    <Modal
-      onClose={() => setErrorModalOpen(false)}
-      open={errorModalOpen}
-      closeIcon
-    >
-      <Modal.Header>Network Error</Modal.Header>
-      <Modal.Content>
-        <Modal.Description>
-          <p>Seems like logout API call resulted with CORS error.</p>
-          <p>
-            You may need to add your origin
-            {' '}
-            {window.location.origin}
-            {' '}
-            to list of trusted origins in your
-            {' '}
-            <a href={corsAdminUrl} target="_blank" rel="noreferrer">Okta Administrator Dashboard</a>
-          </p>
-          <p>
-            Read
-            {' '}
-            <a href={guideUrl} target="_blank" rel="noreferrer">this guide</a>
-            {' '}
-            for more info.
-          </p>
-        </Modal.Description>
-      </Modal.Content>
-    </Modal>
-  );
 
   return (
     <div>
@@ -108,7 +63,6 @@ const Navbar = () => {
           )}
         </Container>
       </Menu>
-      {errorModal}
     </div>
   );
 };
